refactor, move task details mappers
