+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/composable/bottomSheet/task/taskDetailsBottomSheet/DetailsUiState.kt
@@ -30,45 +30,3 @@ data class DetailsUiState(
         .toLocalDateTime(TimeZone.currentSystemDefault()).toString(),
 )
 
-
-fun DetailsUiState.toTaskUiState(): TaskUiState {
-    return TaskUiState(
-        id = this.id,
-        title = this.title,
-        description = this.description,
-        dueDate = this.dueDate,
-        priority = this.priority,
-        status = this.status,
-        categoryId = this.id,
-        createdAt = this.createdAt
-    )
-}
-
-fun DetailsUiState.toTask(): Task {
-    return Task(
-        id = this.id,
-        title = this.title,
-        description = this.description,
-        status = this.status.toTaskStatus(),
-        dueDate = LocalDate.parse(this.dueDate),
-        priority = this.priority.toTaskPriority(),
-        categoryId = this.categoryId,
-        createdAt = LocalDateTime.parse(this.createdAt),
-    )
-}
-
-fun Task.toDetailsState(): DetailsUiState {
-    return DetailsUiState(
-        id = this.id,
-        title = this.title,
-        dueDate = this.dueDate.toString(),
-        categoryId = this.categoryId,
-        priority = this.priority.toState(),
-        description = this.description ?: "",
-        status = this.status.toState(),
-        createdAt = this.createdAt.toString(),
-        categoryImagePath = "",
-        moveStatusToLabel = ""
-    )
-}
-
