+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/composable/bottomSheet/task/taskDetailsBottomSheet/TaskDetailsComponent.kt
@@ -30,6 +30,7 @@ import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
 import com.sanaa.tudee_assistant.presentation.model.TaskUiState
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
+import com.sanaa.tudee_assistant.presentation.model.mapper.toTaskUiState
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
